refactor: apply code readability principles to dashboard and workers
- Split `DashboardPage` into smaller sub-components and extracted data fetching logic.
- Extracted `TransferDialog` and `AffiliationDialog` from `WorkersClient` into separate files.
- Refactored worker server actions into smaller, single-purpose helper functions.
- Centralized common types in `app/(app)/masters/workers/types.ts`.
- Improved JSDoc documentation and formatting across refactored files.
- Ensured compliance with 'Clean Code' principles and project guidelines.
