+++ app/(app)/masters/workers/workers-client.tsx
@@ -3,43 +3,16 @@ import { useMemo, useState, useCallback } from 'react';
 import { useRouter } from 'next/navigation';
 import { createColumnHelper } from '@tanstack/react-table';
 import type { ColumnDef } from '@tanstack/react-table';
-import { toast } from 'sonner';
 import { Button } from '@/components/ui/button';
 import { Sheet, SheetContent, SheetHeader, SheetTitle } from '@/components/ui/sheet';
-import {
-  Dialog,
-  DialogContent,
-  DialogHeader,
-  DialogTitle,
-  DialogFooter,
-} from '@/components/ui/dialog';
-import { Input } from '@/components/ui/input';
-import { Label } from '@/components/ui/label';
 import { MasterShell } from '@/components/master-shell';
 import { DataGrid } from '@/components/data-grid';
 import { EmptyState } from '@/components/empty-state';
 import { useOptimalPageSize } from '@/lib/hooks/use-optimal-page-size';
-import { SearchableSelect } from '@/components/searchable-select';
 import { WorkerForm } from './worker-form';
-import { transferWorker, changeAffiliation } from './actions';
-import { EMPLOYMENT_TYPES } from '@/lib/validators/worker';
-
-type Worker = {
-  id: string;
-  code: string;
-  full_name: string;
-  phone: string | null;
-  home_city: string | null;
-  is_active: boolean;
-  current_site_id: string;
-  site_code: string | null;
-  site_name: string | null;
-  employment_type: string | null;
-  contractor_name: string | null;
-};
-
-type SiteOption = { id: string; name: string; code: string };
-type PartyOption = { id: string; name: string; short_code: string | null; type: string };
+import { TransferDialog } from './transfer-dialog';
+import { AffiliationDialog } from './affiliation-dialog';
+import type { Worker, SiteOption, PartyOption } from './types';
 
 type Props = {
   workers: Worker[];
@@ -86,12 +59,6 @@ const EXPORT_COLS = [
   { key: 'phone' as const, header: 'Phone' },
   { key: 'home_city' as const, header: 'Home city' },
 ];
-
-const EMPLOYMENT_LABELS: Record<(typeof EMPLOYMENT_TYPES)[number], string> = {
-  DIRECT: 'Direct',
-  CONTRACTOR_EMPLOYEE: 'Contractor employee',
-  SUBCONTRACTOR_LENT: 'Sub-contractor (lent)',
-};
 
 export function WorkersClient({ workers, sites, parties }: Props) {
   const router = useRouter();
@@ -245,215 +212,5 @@ export function WorkersClient({ workers, sites, parties }: Props) {
         }}
       />
     </div>
-  );
-}
-
-type DlgProps<T> = {
-  worker: Worker | null;
-  onClose: () => void;
-  onDone: () => void;
-} & T;
-
-function TransferDialog({ worker, sites, onClose, onDone }: DlgProps<{ sites: SiteOption[] }>) {
-  const today = new Date().toISOString().slice(0, 10);
-  const [toSiteId, setToSiteId] = useState<string | null>(null);
-  const [effectiveFrom, setEffectiveFrom] = useState(today);
-  const [reason, setReason] = useState('');
-  const [busy, setBusy] = useState(false);
-
-  const open = worker !== null;
-  const otherSites = useMemo(
-    () => sites.filter((s) => s.id !== worker?.current_site_id),
-    [sites, worker],
-  );
-
-  async function submit() {
-    if (!worker || !toSiteId || !reason.trim()) return;
-    setBusy(true);
-    try {
-      const res = await transferWorker({
-        worker_id: worker.id,
-        to_site_id: toSiteId,
-        effective_from: effectiveFrom,
-        reason: reason.trim(),
-      });
-      if (!res.ok) {
-        toast.error(res.error);
-        return;
-      }
-      toast.success('Worker transferred.');
-      onDone();
-      setToSiteId(null);
-      setEffectiveFrom(today);
-      setReason('');
-    } finally {
-      setBusy(false);
-    }
-  }
-
-  return (
-    <Dialog open={open} onOpenChange={(v) => (v ? null : onClose())}>
-      <DialogContent>
-        <DialogHeader>
-          <DialogTitle>Transfer worker</DialogTitle>
-        </DialogHeader>
-        {worker && (
-          <div className="grid gap-3">
-            <p className="text-muted-foreground text-sm">
-              Moving <span className="font-medium">{worker.full_name}</span> ({worker.code}) from{' '}
-              <span className="font-medium">{worker.site_code ?? '—'}</span>.
-            </p>
-            <div className="grid gap-1.5">
-              <Label>To site *</Label>
-              <SearchableSelect
-                options={otherSites.map((s) => ({
-                  value: s.id,
-                  label: `${s.code} — ${s.name}`,
-                }))}
-                value={toSiteId}
-                onChange={setToSiteId}
-                placeholder="Select site"
-              />
-            </div>
-            <div className="grid gap-1.5">
-              <Label htmlFor="t-eff-from">Effective from *</Label>
-              <Input
-                id="t-eff-from"
-                type="date"
-                value={effectiveFrom}
-                onChange={(e) => setEffectiveFrom(e.target.value)}
-              />
-            </div>
-            <div className="grid gap-1.5">
-              <Label htmlFor="t-reason">Reason *</Label>
-              <Input
-                id="t-reason"
-                value={reason}
-                onChange={(e) => setReason(e.target.value)}
-                placeholder="Why is this worker being transferred?"
-                maxLength={200}
-              />
-            </div>
-          </div>
-        )}
-        <DialogFooter>
-          <Button variant="outline" onClick={onClose} disabled={busy}>
-            Cancel
-          </Button>
-          <Button onClick={submit} disabled={!toSiteId || !reason.trim() || busy}>
-            {busy ? 'Transferring…' : 'Transfer'}
-          </Button>
-        </DialogFooter>
-      </DialogContent>
-    </Dialog>
-  );
-}
-
-function AffiliationDialog({
-  worker,
-  parties,
-  onClose,
-  onDone,
-}: DlgProps<{ parties: PartyOption[] }>) {
-  const today = new Date().toISOString().slice(0, 10);
-  const [type, setType] = useState<(typeof EMPLOYMENT_TYPES)[number] | null>(null);
-  const [partyId, setPartyId] = useState<string | null>(null);
-  const [effectiveFrom, setEffectiveFrom] = useState(today);
-  const [busy, setBusy] = useState(false);
-
-  const open = worker !== null;
-  const needsParty = type !== null && type !== 'DIRECT';
-
-  async function submit() {
-    if (!worker || !type) return;
-    if (needsParty && !partyId) return;
-    setBusy(true);
-    try {
-      const res = await changeAffiliation({
-        worker_id: worker.id,
-        employment_type: type,
-        contractor_party_id: type === 'DIRECT' ? null : partyId,
-        effective_from: effectiveFrom,
-      });
-      if (!res.ok) {
-        toast.error(res.error);
-        return;
-      }
-      toast.success('Affiliation updated.');
-      onDone();
-      setType(null);
-      setPartyId(null);
-      setEffectiveFrom(today);
-    } finally {
-      setBusy(false);
-    }
-  }
-
-  const typeOptions = EMPLOYMENT_TYPES.map((t) => ({
-    value: t,
-    label: EMPLOYMENT_LABELS[t],
-  }));
-  const partyOptions = parties.map((p) => ({
-    value: p.id,
-    label: p.name,
-    sub: p.short_code ?? p.type,
-  }));
-
-  return (
-    <Dialog open={open} onOpenChange={(v) => (v ? null : onClose())}>
-      <DialogContent>
-        <DialogHeader>
-          <DialogTitle>Change affiliation</DialogTitle>
-        </DialogHeader>
-        {worker && (
-          <div className="grid gap-3">
-            <p className="text-muted-foreground text-sm">
-              Updating employment for <span className="font-medium">{worker.full_name}</span> (
-              {worker.code}). Current: {worker.employment_type ?? '—'}.
-            </p>
-            <div className="grid gap-1.5">
-              <Label>New type *</Label>
-              <SearchableSelect
-                options={typeOptions}
-                value={type}
-                onChange={(v) => {
-                  setType(v);
-                  if (v === 'DIRECT') setPartyId(null);
-                }}
-                placeholder="Select type"
-              />
-            </div>
-            {needsParty && (
-              <div className="grid gap-1.5">
-                <Label>Contractor *</Label>
-                <SearchableSelect
-                  options={partyOptions}
-                  value={partyId}
-                  onChange={setPartyId}
-                  placeholder="Select contractor"
-                />
-              </div>
-            )}
-            <div className="grid gap-1.5">
-              <Label htmlFor="a-eff-from">Effective from *</Label>
-              <Input
-                id="a-eff-from"
-                type="date"
-                value={effectiveFrom}
-                onChange={(e) => setEffectiveFrom(e.target.value)}
-              />
-            </div>
-          </div>
-        )}
-        <DialogFooter>
-          <Button variant="outline" onClick={onClose} disabled={busy}>
-            Cancel
-          </Button>
-          <Button onClick={submit} disabled={!type || (needsParty && !partyId) || busy}>
-            {busy ? 'Saving…' : 'Update'}
-          </Button>
-        </DialogFooter>
-      </DialogContent>
-    </Dialog>
   );
 }
